// app/components/Wiki/types/PageSearch.ts
// Generated by https://quicktype.io

export interface SearchResponse {
	batchcomplete: string;
	query: Query;
	continue?: Continue;
}

export interface Continue {
	sroffset: number;
	continue: string;
}

export interface Query {
	search: Search[];
	searchinfo?: Searchinfo;
}

export interface Search {
	ns: number;
	title: string;
	pageid: number;
	size: number;
	wordcount: number;
	snippet: string;
	timestamp: string;
}

export interface Searchinfo {
	totalhits: number;
}